rename: chain.py → skill_chain.py
Aligns the module name with the class name (SkillChainDocument).
