+++ tests/test_lint.py
@@ -180,7 +180,7 @@ class TestChainAutoDetection:
         root.mkdir()
 
         with _patch("wos.project.validate_project", return_value=[]), \
-             _patch("wos.chain.validate_chain") as mock_chain:
+             _patch("wos.skill_chain.validate_chain") as mock_chain:
             _run_audit("--root", str(root), "--no-urls")
 
         mock_chain.assert_not_called()
@@ -211,7 +211,7 @@ class TestChainAutoDetection:
         self._write_chain_manifest(hidden / "nested.chain.md", goal="some goal")
 
         with _patch("wos.project.validate_project", return_value=[]), \
-             _patch("wos.chain.validate_chain") as mock_chain:
+             _patch("wos.skill_chain.validate_chain") as mock_chain:
             _run_audit("--root", str(root), "--no-urls")
 
         mock_chain.assert_not_called()
