+++ wos/__init__.py
@@ -1,7 +1,7 @@
 """wos: Claude Code plugin for structured project context."""
 # Import type modules so subclasses self-register via @Document.register
-import wos.chain  # noqa: F401
 import wos.plan  # noqa: F401
 import wos.research  # noqa: F401
 import wos.skill  # noqa: F401
+import wos.skill_chain  # noqa: F401
 import wos.wiki  # noqa: F401
